Optimise updating RecyclerView adapters
Previously RecyclerView adapters were updated by calling notifyDataSetChanged(), which forces the layout manager to rebind and relayout every item in the RecyclerView, no matter if they were changed or not. Now the adapter is updated using DiffUtil, which calculates the changes and defines an optimal set of update-operations to migrate from the old items list to the new one.