Fix adapter not updating after deleting grocery item
This was because the grocery list was is only refreshed when MainActivity is
recreated. In ConfigurationsActivity this doesn't happen, because there swiping
causes a direct deletion from FoodItemViewModel and the underlying LiveData.

In MainActivity the deletion only add the item to the list of removed items
which is used to filter the grocery list when the activity is recreated, so
changes are not reflected immediately.